Prepare the output of r2dii.analysis::target_sda() for plot_timeline()

Prepare the output of r2dii.analysis::target_sda() for plot_timeline()

prep_timelineY(
  data,
  value = "emission_factor_value",
  metric = "emission_factor_metric",
  sector_filter = NULL,
  extrapolate = FALSE
)

Arguments

data
Data frame like the output of r2dii.analysis::target_sda().

value
String of length 1. The name of the column holding the value to plot.

metric
String of length 1. The name of the column holding the metrics to plot.

sector_filter
String of length 1. Sector to pick from the data.

extrapolate
Logical of length 1. TRUE extrapolates to match the furthest value in the data set.

Value

Data frame with columns: year, line_name, value, extrapolated.

See also

Examples

library(dplyr, warn.conflicts = FALSE)

# Fails
try(prep_timelineY(sda))

#> Error : `sector` must have a single value. It has: automotive, aviation, cement, oil and gas, shipping, coal, steel.

# Works: meets `data` requirements
prep_timelineY(sda, sector_filter = "cement")

#> # A tibble: 43 x 7
#>    sector year       emission_factor_metric emission_factor_val… line_name value
#>    <chr>  <date>     <chr>                                 <dbl> <chr>     <dbl>
#>  1 cement 2013-01-01 projected                             0.658 projected 0.658
#>  2 cement 2014-01-01 projected                             0.659 projected 0.659
#>  3 cement 2015-01-01 projected                             0.660 projected 0.660
#>  4 cement 2016-01-01 projected                             0.661 projected 0.661
#>  5 cement 2017-01-01 projected                             0.662 projected 0.662
#>  6 cement 2018-01-01 projected                             0.662 projected 0.662
#>  7 cement 2019-01-01 projected                             0.663 projected 0.663
#>  8 cement 2020-01-01 projected                             0.664 projected 0.664
#>  9 cement 2021-01-01 projected                             0.665 projected 0.665
#> 10 cement 2022-01-01 projected                             0.666 projected 0.666
#> # … with 33 more rows, and 1 more variable: extrapolated <lgl>

# Same
sda %>%
  filter(sector == "cement") %>%
  prep_timelineY()

#> # A tibble: 43 x 7
#>    sector year       emission_factor_metric emission_factor_val… line_name value
#>    <chr>  <date>     <chr>                                 <dbl> <chr>     <dbl>
#>  1 cement 2013-01-01 projected                             0.658 projected 0.658
#>  2 cement 2014-01-01 projected                             0.659 projected 0.659
#>  3 cement 2015-01-01 projected                             0.660 projected 0.660
#>  4 cement 2016-01-01 projected                             0.661 projected 0.661
#>  5 cement 2017-01-01 projected                             0.662 projected 0.662
#>  6 cement 2018-01-01 projected                             0.662 projected 0.662
#>  7 cement 2019-01-01 projected                             0.663 projected 0.663
#>  8 cement 2020-01-01 projected                             0.664 projected 0.664
#>  9 cement 2021-01-01 projected                             0.665 projected 0.665
#> 10 cement 2022-01-01 projected                             0.666 projected 0.666
#> # … with 33 more rows, and 1 more variable: extrapolated <lgl>

prep_timelineY(sda, sector_filter = "cement", extrapolate = TRUE)

#> # A tibble: 49 x 7
#>    sector year       emission_factor_metric emission_factor_val… line_name value
#>    <chr>  <date>     <chr>                                 <dbl> <chr>     <dbl>
#>  1 cement 2013-01-01 projected                             0.658 projected 0.658
#>  2 cement 2014-01-01 projected                             0.659 projected 0.659
#>  3 cement 2015-01-01 projected                             0.660 projected 0.660
#>  4 cement 2016-01-01 projected                             0.661 projected 0.661
#>  5 cement 2017-01-01 projected                             0.662 projected 0.662
#>  6 cement 2018-01-01 projected                             0.662 projected 0.662
#>  7 cement 2019-01-01 projected                             0.663 projected 0.663
#>  8 cement 2020-01-01 projected                             0.664 projected 0.664
#>  9 cement 2021-01-01 projected                             0.665 projected 0.665
#> 10 cement 2022-01-01 projected                             0.666 projected 0.666
#> # … with 39 more rows, and 1 more variable: extrapolated <lgl>

data <- sda %>%
  filter(sector == "cement") %>%
  rename(
    custom_value = "emission_factor_value",
    custom_metric = "emission_factor_metric"
  )
prep_timelineY(data, value = "custom_value", metric = "custom_metric")

#> # A tibble: 43 x 7
#>    sector year       custom_metric custom_value line_name value extrapolated
#>    <chr>  <date>     <chr>                <dbl> <chr>     <dbl> <lgl>       
#>  1 cement 2013-01-01 projected            0.658 projected 0.658 FALSE       
#>  2 cement 2014-01-01 projected            0.659 projected 0.659 FALSE       
#>  3 cement 2015-01-01 projected            0.660 projected 0.660 FALSE       
#>  4 cement 2016-01-01 projected            0.661 projected 0.661 FALSE       
#>  5 cement 2017-01-01 projected            0.662 projected 0.662 FALSE       
#>  6 cement 2018-01-01 projected            0.662 projected 0.662 FALSE       
#>  7 cement 2019-01-01 projected            0.663 projected 0.663 FALSE       
#>  8 cement 2020-01-01 projected            0.664 projected 0.664 FALSE       
#>  9 cement 2021-01-01 projected            0.665 projected 0.665 FALSE       
#> 10 cement 2022-01-01 projected            0.666 projected 0.666 FALSE       
#> # … with 33 more rows