Prepares pre-processed data for plotting a trajectory chart

Prepares pre-processed data for plotting a trajectory chart

prep_trajectory(
  data,
  sector_filter,
  technology_filter,
  region_filter,
  scenario_source_filter,
  value_name,
  end_year_filter = 2025,
  normalize_to_start_year = TRUE
)

prep_trajectoryB(data, value = "production", normalize = TRUE)

Arguments

data
Pre-processed input data.

sector_filter
Sector for which to filter the data (character string).

technology_filter
Technology for which to filter the data (character string).

region_filter
Region for which to filter the data (character string).

scenario_source_filter
Scenario source for which to filter the data (character string).

value_name, value
The name of the value to be plotted in the trajectory chart (character string).

end_year_filter
Cut-off year for the chart (an integer).

normalize_to_start_year, normalize
Logical of length-1. TRUE normalized to start year.

Value

A data frame.

Examples

prep_trajectory(
  market_share,
  sector_filter = "power",
  technology_filter = "oilcap",
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  value_name = "production"
)

#> # A tibble: 30 x 5
#>     year metric_type metric            technology value
#>    <int> <chr>       <chr>             <chr>      <dbl>
#>  1  2020 portfolio   projected         oilcap     1    
#>  2  2020 benchmark   corporate_economy oilcap     1    
#>  3  2020 scenario    cps               oilcap     1    
#>  4  2020 scenario    sds               oilcap     1    
#>  5  2020 scenario    sps               oilcap     1    
#>  6  2021 portfolio   projected         oilcap     0.994
#>  7  2021 benchmark   corporate_economy oilcap     0.992
#>  8  2021 scenario    cps               oilcap     0.932
#>  9  2021 scenario    sds               oilcap     0.922
#> 10  2021 scenario    sps               oilcap     0.931
#> # … with 20 more rows

library(dplyr)

data <- market_share %>%
  filter(
    technology == "oilcap",
    region == "global",
    scenario_source == "demo_2020",
    year <= 2025,
    sector == "power"
  )

prep_trajectoryB(data)

#> # A tibble: 30 x 5
#>     year metric_type metric            technology value
#>    <int> <chr>       <chr>             <chr>      <dbl>
#>  1  2020 portfolio   projected         oilcap     1    
#>  2  2020 benchmark   corporate_economy oilcap     1    
#>  3  2020 scenario    cps               oilcap     1    
#>  4  2020 scenario    sds               oilcap     1    
#>  5  2020 scenario    sps               oilcap     1    
#>  6  2021 portfolio   projected         oilcap     0.994
#>  7  2021 benchmark   corporate_economy oilcap     0.992
#>  8  2021 scenario    cps               oilcap     0.932
#>  9  2021 scenario    sds               oilcap     0.922
#> 10  2021 scenario    sps               oilcap     0.931
#> # … with 20 more rows