Prepares pre-processed data for plotting a tech-mix chart

Prepares pre-processed data for plotting a tech-mix chart

prepare_for_techmix_chart(
  data,
  sector_filter = c("automotive", "aviation", "cement", "oil and gas", "shipping",
    "steel", "power"),
  years_filter = NULL,
  region_filter = "global",
  scenario_source_filter = NULL,
  scenario_filter = NULL,
  value_to_plot = "technology_share"
)

Arguments

data
Pre-processed input data.

sector_filter
Sector for which to filter the data (character string).

years_filter
Years to plot in the graph (array of integer values).

region_filter
Region for which to filter the data (character string).

scenario_source_filter
Scenario source for which to filter the data (character string).

scenario_filter
Scenario to plot in the graph (character string).

value_to_plot
The name of the value to be plotted as a bar chart (character string).

Examples

prepare_for_techmix_chart(
  process_input_data(example_data),
  sector_filter = "power",
  years_filter = c(2020, 2025),
  region_filter = "global",
  scenario_source_filter = "demo_2020",
  scenario_filter = "sds",
  value_to_plot = "technology_share"
)

#> # A tibble: 30 x 6
#>    sector technology metric_type    metric             value scenario_source
#>    <chr>  <chr>      <chr>          <chr>              <dbl> <chr>          
#>  1 power  coalcap    portfolio_2020 projected         0.129  demo_2020      
#>  2 power  coalcap    benchmark_2020 corporate_economy 0.374  demo_2020      
#>  3 power  coalcap    scenario_2020  sds               0.384  demo_2020      
#>  4 power  coalcap    portfolio_2025 projected         0.117  demo_2020      
#>  5 power  coalcap    benchmark_2025 corporate_economy 0.374  demo_2020      
#>  6 power  coalcap    scenario_2025  sds               0.316  demo_2020      
#>  7 power  gascap     portfolio_2020 projected         0.0999 demo_2020      
#>  8 power  gascap     benchmark_2020 corporate_economy 0.118  demo_2020      
#>  9 power  gascap     scenario_2020  sds               0.0745 demo_2020      
#> 10 power  gascap     portfolio_2025 projected         0.0885 demo_2020      
#> # … with 20 more rows