Prepares sda-type data for timeline plot

Prepares sda-type data for timeline plot

prepare_for_timelineA(
  sda_data,
  sector_filter = c("automotive", "aviation", "cement", "oil and gas", "shipping",
    "steel", "power"),
  year_start = 0,
  year_end = Inf,
  column_line_names = "emission_factor_metric",
  value_to_plot = "emission_factor_value",
  extrapolate_missing_values = FALSE
)

prepare_for_timelineB(data, extrapolate = FALSE)

Arguments

sector_filter
Sector to be used for filtering (character string of length 1). One of automotive, aviation, cement, oil and gas, shipping, steel, power)```.

year_start
Start year of the plot (double).

year_end
End year of the plot (double).

column_line_names
Column specifying the names of lines to be plotted (character string).

value_to_plot
Column name of the value to be plotted (character string).

data, sda_data
Dataframe with columns sector, year and two other columns specifying value to be be plotted as timelines and line names (dataframe).

extrapolate, extrapolate_missing_values
Flag indicating if values should be extrapolated to match the furthest value in the data set.

Value

Dataframe with columns: year, line_name, value, extrapolated.

Examples

# prepare_for_timelineA() --------------------------------------------------

data <- sda
prepare_for_timelineA(
  data,
  sector_filter = "cement",
  year_start = 2020,
  year_end = 2050, column_line_names = "emission_factor_metric",
  value_to_plot = "emission_factor_value",
  extrapolate_missing_values = TRUE
)

#> # A tibble: 35 x 5
#>    year       line_name         value extrapolated sector
#>    <date>     <chr>             <dbl> <lgl>        <chr> 
#>  1 2020-01-01 projected         0.664 FALSE        cement
#>  2 2021-01-01 projected         0.665 FALSE        cement
#>  3 2022-01-01 projected         0.666 FALSE        cement
#>  4 2023-01-01 projected         0.667 FALSE        cement
#>  5 2024-01-01 projected         0.668 FALSE        cement
#>  6 2025-01-01 projected         0.669 FALSE        cement
#>  7 2020-01-01 corporate_economy 0.669 FALSE        cement
#>  8 2021-01-01 corporate_economy 0.670 FALSE        cement
#>  9 2022-01-01 corporate_economy 0.671 FALSE        cement
#> 10 2023-01-01 corporate_economy 0.672 FALSE        cement
#> # … with 25 more rows

# prepare_for_timelineA() --------------------------------------------------

data <- sda
tail(prepare_for_timelineB(data))

#> # A tibble: 6 x 5
#>   year       line_name               value extrapolated sector
#>   <date>     <chr>                   <dbl> <lgl>        <chr> 
#> 1 2040-01-01 adjusted_scenario_demo 0.236  FALSE        cement
#> 2 2040-01-01 adjusted_scenario_demo 0.0654 FALSE        steel 
#> 3 2045-01-01 adjusted_scenario_demo 0.189  FALSE        cement
#> 4 2045-01-01 adjusted_scenario_demo 0.0327 FALSE        steel 
#> 5 2050-01-01 adjusted_scenario_demo 0.0945 FALSE        cement
#> 6 2050-01-01 adjusted_scenario_demo 0.0262 FALSE        steel 

tail(prepare_for_timelineB(data, extrapolate = TRUE))

#> # A tibble: 6 x 5
#>   year       line_name         value extrapolated sector    
#>   <date>     <chr>             <dbl> <lgl>        <chr>     
#> 1 2026-01-01 corporate_economy 0.279 TRUE         automotive
#> 2 2026-01-01 projected         0.178 TRUE         automotive
#> 3 2025-01-01 target_demo       0.521 TRUE         cement    
#> 4 2050-01-01 corporate_economy 0.279 TRUE         automotive
#> 5 2050-01-01 projected         0.178 TRUE         automotive
#> 6 2050-01-01 target_demo       0.521 TRUE         cement