r2dii.plot: Visualize the Climate Scenario Alignment of a Financial Portfolio

Create plots to visualize the alignment of a corporate lending financial portfolio to climate change scenarios based on climate indicators (production and emission intensities) across key climate relevant sectors of the 'PACTA' methodology (Paris Agreement Capital Transition Assessment; <https://2degrees-investing.org/>). Financial institutions use 'PACTA' to study how their capital allocation decisions align with climate change mitigation goals.

See also

Author

Maintainer: Monika Furdyna monika.furdyna@gmail.com (ORCID) [contractor]

Authors:

Mauro Lepore maurolepore@gmail.com (ORCID) [contractor]

Alex Axthelm alex@2degrees-investing.org (ORCID) [contractor]

Other contributors:

2 Degrees Investing Initiative contact@2degrees-investing.org [copyright holder, funder]